Implement blob validator
